Connect bottom row sites to their neighbors
While top row sites need only be connected to the virtual top site,
bottom row sites must be connected somehow to their neighbors, specifically
when there is a path to the second-from-bottom row, and a bottom
row site is opened.  If that new site does not check its neighbors,
then although a path exists, none will be recorded.