Also have new top-row sites check neighbors
By the same logic as is used for having the bottom-row sites check
their neighbors, we need to have the top-row sites check their neighbors.

Specifically, if the top row site is the last one added, it needs
to union with something to its left or right or below, in order to
complete the recorded path.  Otherwise, a path will exist, but will
not be known.